Check value equality when determining stage completion
The current 'is-level-complete' code does not check if the values
of two expressions are equal, only if they have the same tree of
nodes. This leads to problems for expressions that take on
different values, but have the same node tree, e.g. NumberExpr:
any number can stand in for any other number. This patch changes
the check to include making sure that `isValue` and `value` are
both equal if present.