Fix flickering after choosing level
EnvironmentDisplay#update was setting the global stage to its own
stage. When transitioning from a level to the chapter select, this
caused the global stage to be overwritten, leading to the wrong
stage being invalidated when the user then chose a new level. Thus,
both stages would draw, causing flickering.