Fix highlight of ES6 lambda hole being in wrong position
clone() shallow-copies properties that are not explicitly children,
leading to an edge case where a duplicated lambda would have an
argument expression that retained the parent of the original,
pre-duplication lambda, causing positioning to be incorrect. To
avoid this, we use getters to access the expression in the holes
array, rather than setting an attribute on the object directly.