Fix how clone() works with VarExpr
Properties that are directly assigned to VarExpr in the constructor
(this.name = foo…) get 1-1 copied without recursive cloning when
the overall VarExpr is cloned. This means the parent-child
relationship can get broken, leading to incorrect positioning of
objects. (To see this: x => y, x => x x x. Drag the former into the
latter, then try to apply one of the resulting expressions. This
will create the '?' animation, but in the wrong place.)